+++ src/private/release/WriteNovaResolvedLocalPublishTarget.ps1
@@ -1,0 +1,13 @@
+function Write-NovaResolvedLocalPublishTarget {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][pscustomobject]$PublishInvocation
+    )
+
+    if (-not $PublishInvocation.IsLocal) {
+        return
+    }
+
+    Write-Verbose "Using $( $PublishInvocation.Target ) as path"
+}
+